requirements.txt
anyio==3.3.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
atpublic==1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
black==21.7b0
bleach==3.3.1
boto3==1.18.6
botocore==1.21.6
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==3.4.7
cycler==0.10.0
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.5.0
dpath==2.0.1
dulwich==0.20.24
dvc==1.11.16
entrypoints==0.3
fastapi==0.67.0
flake8==3.9.2
flatten-dict==0.3.0
flufl.lock==3.2
fsspec==2021.7.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.7
gitdb2==4.0.2
GitPython==3.1.19
grandalf==0.6
h11==0.12.0
idna==3.1
importlib-metadata==4.6.1
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonpath-ng==1.5.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
mailchecker==4.0.11
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.8.0
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.1
notebook==6.4.0
numpy==1.21.1
olefile==0.46
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.2
parso==0.8.2
pathlib2==2.3.6
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.28
pickleshare==0.7.5
Pillow==8.3.1
pip==21.1.3
pluggy==0.13.1
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydot==1.4.2
pyflakes==2.3.1
pygit2==1.6.1
Pygments==2.9.0
pygtrie==2.3.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
pytest-mock==3.6.1
python-benedict==0.24.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.26.0
requests-unixsocket==0.2.0
rich==10.6.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
setuptools==49.6.0.post20210108
shortuuid==1.0.1
shtab==1.3.9
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
starlette==0.14.2
tabulate==0.8.9
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
uvicorn==0.14.0
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wheel==0.36.2
widgetsnbextension==3.5.1
xmltodict==0.12.0
zc.lockfile==2.0
zipp==3.5.0
